backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==1.0.3  # Downgrade to stable version
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.24.1  # Add specific httpx version